freebsd-src: change ninja target names e.g. freebsd.build.14.2-RELEASE
This adopts a new pattern, based on better understanding of ninja.
ninja hides any targets that are inputs to other targets.
This introduces the mod.* pattern, where mod.* are phony targets
that can serve as inputs to other targets. phony targets intended
to be called from the command line depend on these.