freebsd-src: Build src in a temp dir; consolidate build+release
The FreeBSD build scripts relies on being built in a git repo. This
works okay with jj when building from main - but when it's a
workspace, there is no git dir. The git tools are unable to locate git
information, and read info from the top-level lab repo instead.

This does some gymnastics to create a tag based on the jj head, then
hard reset to that tag in a temp dir. I combined build+release because
the jj HEAD tag may not be consistent between runs.

It's a definite workaround the fact that jj doesn't provide a git dir
to workspaces... yet.

Also: Rename configurations afer their major versions