katai/bowling: (abandoned) Initial attempt at prompt engineering an example score
This is really my first attempt at using an LLM in this style.
I was very focused on getting the LLM to produce the correct results.
I kept tweaking the inputs, but I never got there.

My take is that I was biting off more than I and/or the LLM could chew.

---

My experience so far:

- It's very easy to get in the ballpark.
- It's quite a bit harder to refine it to a correct result.
- Examples seem to help quite a bit.
- I'm never quite sure which part of the prompt influences which part of the result
- I find myself wondering if I need to revise the prompt, or try a more intelligent model.

One idea is to start with a fairly sophisticated model, get the
correct result, and then downshift to a less sophisticated model.