katai/bowling: Implement Complete spare frame
This was a bit of an ordeal. LLM started by hardcoding a value. I
tried to get it to change, but it then tried to identify frame
position by value. This obviously doesn't work because two frames can
have identical values.

Interestingly, it then tried storing the list of frames using
Process.put/2 and Process.get/1.

It seemed to really struggle with the idea that it could find the next
item in a list pretty simply. My guess is that the bulk of training
data for the LLM is not immutable functional code. Maybe that's
something I can add to the prompt?