katai/bowling: Refactor to compute scores considering bonus
I did this refactor myself. I wasn't able to get LLM to do anything
approaching it. Maybe if I were more instructive with the comments.

I think this is quite a bit cleaner. It more clearly illustrates the
algorithm of incorporating bonuses into a frame.

It has a special case for the first frame being an open spare, where
there is no score yet to calculate.