katai/bowling: Handle tenth frame
Again, LLM was not particularly helpful here. The one thing it did was
recognize how I wrote something, and reuse the pattern.

When I first instructed LLM to convert the "perfect game" test, it
produced a list of 12 strikes, as expected. I then manually converted
it to use List.duplicate/2 and Enum.reduce/3. Later instructions to
convert tests resulted in it reusing this duplicate / reduce pattern,
rather than producing a long list.