=== DEV ===
This is an experiment in using jujutsu to facilitate development. One of
its unique features is the ability to manipulate the git DAG easily. I
can have a single dev commit that has multiple parents - a mega merge
commit - and then push commits anywhere in the DAG between trunk and
dev.

I am concerned that I will have too much WIP. But that's the challenge
with too much WIP in the first place. All this does is make it visible.