emacs: prescient.el and savehist for more ergonomic completion
Vertico provides the completion, but out of the box it's a pain to use
because there are so many entries. The one I want is usually buried
somewhere. savehist keeps track of what I choose. prescient.el
analyzes the history to determine which is the most likely one based
on frequency and recency.

Change-Id: Ia906101fa40616d8dcb7ef4c6206f89e6a6a6964